+++ apps/bare-expo/e2e/README.md
@@ -3,7 +3,7 @@
 ### Prerequisites
 
 - Install [Maestro CLI](https://maestro.mobile.dev/docs/getting-started/installation)
-- (optional, recommended) Alignment with devices which are used in CI ([iOS](https://github.com/expo/expo/blob/051a306ce7c5b875f7398450e5aeec2e52e313ae/apps/bare-expo/scripts/start-ios-e2e-test.ts#L18), [Android](https://github.com/expo/expo/blob/051a306ce7c5b875f7398450e5aeec2e52e313ae/.github/actions/use-android-emulator/action.yml#L48)). This is necessary for assertions on what is visible on the screen and (especially) for screenshots to match.
+- (optional, recommended) Alignment with devices which are used in CI ([iOS](https://github.com/expo/expo/blob/051a306ce7c5b875f7398450e5aeec2e52e313ae/apps/bare-expo/scripts/start-ios-e2e-test.ts#L18), [Android](https://github.com/expo/expo/blob/051a306ce7c5b875f7398450e5aeec2e52e313ae/.github/actions/use-android-emulator/action.yml#L48)). This is necessary for assertions on what is visible on the screen and (especially) for view shots to match.
 - use the following command to generate the Android emulator:
 
 ```bash
@@ -18,7 +18,7 @@ avdmanager create avd --force -n pixel_7_pro --package 'system-images;android-36
 
 3. Take a screen in NCL, adjust it for your e2e test needs (you can use `KeyValueBox` component to render values you want to test and use the "copy maestro assertions" button to get yaml that you can paste into your test file).
 
-4. Start the screenshot server if you want to take and compare screenshots:
+4. Start the image comparison server if you want to take and compare view shots:
 
 ```bash
 cd e2e/_nested-flows && bun --watch --no-clear-screen ./image-comparison-server.ts
@@ -32,13 +32,13 @@ cd e2e/_nested-flows && bun --watch --no-clear-screen ./image-comparison-server.
 - Don't assert on looong visible text. Example: when you're using `assertVisible` with a multiline string, that assertion will be hard to maintain.
 - Consider both platforms early on. It can save time later.
 - Limit interactivity. It's not great to say this, but every tapping of a button or scrolling takes time and reduces reliability. The best tests are the ones that deep link to a screen and then do a few assertions.
-- Use screenshots sparingly. They are invaluable for visual verification, but they add to test time and need to be updated when the UI or device changes. View shots "focus" on a specific view and can work cross-platform.
+- View shots "focus" on a specific view and can work cross-platform (view shots taken on different platforms can be normalized to allow their comparison even though they have different sizes). They are the main tool for visual verification, but they add to test time and may need to be updated when the UI, OS or device changes, so don't overuse them.
 - Use `maestro hierarchy` for debugging selectors and understanding what maestro sees on the screen.
 - Ensure every test takes care of starting from the state it expects to start from, and can run independently of others (no reliance on state left by a previous test).
 
 ### Running tests
 
-You can run tests directly from the studio but there's a gotcha with screenshots — studio stores them at a wrong location (different from when you run from CLI). So for screenshots specifically, it's better to leave them to the end when you want to create them for a PR (just use the screenshot feature of the emulator).
+You can run tests directly from the studio but there's a gotcha with view shots — studio stores them at a wrong location (different from when you run from CLI). So for view shots specifically, it's better to leave them to the end when you want to create them for a PR.
 
 To run tests from CLI, use the following command which picks up whatever device you have running (the `platform` param requires [this](https://github.com/mobile-dev-inc/Maestro/commit/12072a4782704dd4a8d2316625e3eb2df8db5bc5)):
 
@@ -57,16 +57,16 @@ maestro test expo-video/test.yaml --shard-all 2
 
 To run the tests the same way they'd run in CI, do `cd apps/bare-expo/scripts` and `./start-ios-e2e-test.ts --test` or `./start-android-e2e-test.ts --test`.
 
-### Comparing screen / view shots
+### Comparing images
 
-Comparison works by always keeping a base image committed in the repository (`.base.png` for cross-platform view shots, or `.base.platform.png` for platform-specific screenshots or view shots). During CI, a new screenshot is taken and compared to the base image. Based on the comparison result, the test assertion will either pass or fail. For examples, see `_nested-flows/screenshot-comparison.yaml` and `_nested-flows/viewshot-comparison.yaml`.
+Comparison works by always keeping a base image committed in the repository (`.base.png` for cross-platform view shots, or `.base.platform.png` for platform-specific view shots). During CI, a new view shot is taken and compared to the base image. Based on the comparison result, the test assertion will either pass or fail. For examples, see use of `_nested-flows/viewshot-comparison.yaml`.
 The shots, as well as diffs against the base images are stored as GH artifacts in CI, and so are the logs of the image comparison server.
 
 ### Troubleshooting
 
-Maestro creates a folder at `./maestro` where you'll find logs and screenshots for when an assertion failed. If a screenshot fails, look at the logs of the screenshot server, it includes path to a screenshot diff.
+Maestro creates a folder at `./maestro` where you'll find logs, failed view shot diffs, and screenshots for when an assertion failed. If a view shot fails, look at the logs of the image comparison server, it includes path to the image diff.
 
-When running in CI, the artifacts (logs, screenshots, screenshot diffs) are uploaded as artifacts and you can download them directly from slack failure message. Also the workflow run logs have a `Artifacts download URL` where you can download them.
+When running in CI, the artifacts (logs, screenshots, view shot diffs) are uploaded as artifacts and you can download them directly from slack failure message. Also the workflow run logs have a `Artifacts download URL` where you can download them.
 
 There's also a maestro slack channel.
