+++ apps/bare-expo/e2e/expo-image/test.yaml
@@ -11,13 +11,13 @@ jsEngine: graaljs
     env:
       screenshotOutputPath: expo-image
       testID: 'header-Appearance'
-      mode: 'crossPlatform'
+      mode: 'normalize'
 - runFlow:
     file: ../_nested-flows/viewshot-comparison.yaml
     env:
       screenshotOutputPath: expo-image
       testID: 'image-comparison-list'
-      mode: 'platformDependent'
+      mode: 'keep-originals'
 
 - runFlow:
     file: ../_nested-flows/screenshot-comparison.yaml
