+++ apps/bare-expo/e2e/image-comparison/src/schema.ts
@@ -4,14 +4,14 @@ const screenshotSchema = z.object({
   baseImage: z.string().nonempty(),
   currentScreenshot: z.string().nonempty(),
   diffOutputPath: z.string().nonempty(),
-  similarityThreshold: z.number().min(0).optional(),
+  similarityThreshold: z.number().min(0).max(100).optional(),
   platform: z.enum(['ios', 'android']),
 });
 
 const viewShotSchema = screenshotSchema.and(
   z.object({
     testID: z.string().nonempty(),
-    mode: z.enum(['crossPlatform', 'platformDependent']),
+    mode: z.enum(['normalize', 'keep-originals']),
   })
 );
