+++ apps/bare-expo/scripts/lib/e2e-common.ts
@@ -177,7 +177,8 @@ const getCustomMaestroFlowsAsync = async (
     ignore.push('**/*.android.yaml');
   }
 
-  const yamlFiles = await glob.glob('**/*.yaml', {
+  // TODO add param to this
+  const yamlFiles = await glob('**/*.yaml', {
     cwd: e2eDir,
     ignore,
   });
